refactoring, https://github.com/PathOnAI/LiteWebAgent/issues/56, https://github.com/PathOnAI/LiteWebAgent/issues/58
