+++ litewebagent/agents/PromptAgents/PromptSearchAgent.py
@@ -1,5 +1,4 @@
 from litellm import completion
-from .BaseAgent import BaseAgent
 from typing import List, Dict, Any
 import json
 import logging
@@ -14,18 +13,19 @@ from litewebagent.observation.observation import (
 )
 from litewebagent.observation.extract_elements import extract_interactive_elements, highlight_elements
 from litewebagent.action.highlevel import HighLevelActionSet
-from litewebagent.playwright_manager import PlaywrightManager
+from litewebagent.utils.playwright_manager import PlaywrightManager
 from litewebagent.action.base import execute_python_code
 from browsergym.utils.obs import flatten_axtree_to_str, flatten_dom_to_str
-from litewebagent.functions.functions import build_highlevel_action_parser
-from litewebagent.functions.utils import *
+from litewebagent.utils.utils import build_highlevel_action_parser
+from litewebagent.utils.utils import *
 import ast
 import pyparsing as pp
 from typing import Any
 from collections import defaultdict
 from collections import deque
-from litewebagent.functions.utils import search_interactive_elements
-from litewebagent.replay import take_action
+from litewebagent.utils.utils import search_interactive_elements
+from litewebagent.utils.replay import take_action
+from litewebagent.utils.prompt_functions import extract_top_actions, is_goal_finished
 
 logger = logging.getLogger(__name__)
 openai_client = OpenAI()
@@ -90,74 +90,14 @@ class PromptSearchAgent:
         for item in trajectory:
             action = item['action']
             messages.append({"role": "user", "content": 'action is: {}'.format(action)})
-        from pydantic import BaseModel
-
-        class Plan(BaseModel):
-            goal_finished: bool
 
 
-        new_response = openai_client.beta.chat.completions.parse(
-            model='gpt-4o-mini',
-            messages=messages,
-            response_format=Plan,
-        )
-        message = new_response.choices[0].message.parsed
-
-        goal_finished = message.goal_finished
+        goal_finished = is_goal_finished(messages, openai_client)
 
         if goal_finished == False:
-
+            updated_actions = extract_top_actions(trajectory, self.goal, page_info, self.action_set, openai_client, branching_factor)
             # Prepare messages for AI model
-            system_msg = f"""
-                    # Instructions
-                    Review the current state of the page and all other information to find the best
-                    possible next action to accomplish your goal. Your answer will be interpreted
-                    and executed by a program, make sure to follow the formatting instructions.
-    
-                    Previous actions and action results are: {trajectory}
-    
-                    Provide ONLY ONE action. Do not suggest multiple actions or a sequence of actions.
-                    # Goal:
-                    {self.goal}"""
-            prompt = prepare_prompt(page_info, self.action_set, 'axtree')
 
-            response = openai_client.chat.completions.create(
-                model="gpt-4o",
-                messages=[
-                    {"role": "system", "content": system_msg},
-                    {"role": "user",
-                     "content": [
-                         {"type": "text", "text": prompt},
-                         {"type": "image_url",
-                          "image_url": {
-                              "url": f"data:image/jpeg;base64,{base64_image}",
-                              "detail": "high"
-                          }
-                          }
-                     ]
-                     },
-                ],
-                n=max(branching_factor * 2, 20)
-            )
-            responses: list[str] = [x.message.content for x in response.choices]
-            highlevel_action_parser = build_highlevel_action_parser()
-            print(responses)
-            parsed_actions_count = defaultdict(int)
-            all_actions = {}
-            for response in responses:
-                result = highlevel_action_parser.parse_string(response)
-                result = result[0] if result else ""  # Convert to string
-                if result not in all_actions:
-                    all_actions[result] = {'action': response}
-                parsed_actions_count[result] += 1
-            print(parsed_actions_count)
-            top_actions = sorted(parsed_actions_count, key=parsed_actions_count.get, reverse=True)[:branching_factor]
-            top_action_count = sum([parsed_actions_count[action] for action in top_actions])
-            updated_actions = []
-            for action in top_actions:
-                a = all_actions[action]
-                a['prob'] = parsed_actions_count[action] / top_action_count
-                updated_actions.append(a)
             return False, updated_actions
         else:
             return True, None
@@ -172,6 +112,7 @@ class PromptSearchAgent:
                 self.trajectories.append({'goal_finished': goal_finished, 'trajectory': trajectory})
                 if not goal_finished:
                     for action in next_actions:
+                        print(action)
                         # Enqueue the new trajectory and increase the depth by 1
                         page = self.playwright_manager.get_page()
                         page_info = extract_page_info(page)
@@ -182,6 +123,10 @@ class PromptSearchAgent:
                                 for function_name, function_args in function_calls:
                                     print(function_name, function_args)
                                     extracted_number = parse_function_args(function_args)
+                                    # TODO: fix the cases where there is no element
+                                    # scroll [0, 200]
+                                    # None
+                                    # An error occurred: 'NoneType' object does not support item assignment
                                     result = search_interactive_elements(page_info["interactive_elements"], extracted_number)
                                     print(result)
                                     result['action'] = action['action']
